수정중 그냥 수정중 -상훈

+++ Thenolja/src/thenolja/tb_hotel/controller/RoomServlet.java
@@ -6,6 +6,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import thenolja.member.model.vo.Member;
 
 /**
  * Servlet implementation class RoomServlet
@@ -27,28 +30,29 @@ public class RoomServlet extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String uri = request.getRequestURI();
-		// 요청한 url 전부 보여줌
-		// System.out.println(uri);
 		
 		String mapping = uri.substring(uri.lastIndexOf("/") + 1, uri.lastIndexOf("."));
-		// System.out.println(mapping);
 		
 		RoomController rc = new RoomController();
+		HttpSession session = request.getSession();
+		Member loginUser = (Member)session.getAttribute("loginUser");
+		
 		String view = "";
 		boolean flag = true;
+		if(loginUser != null && loginUser.getMemStatus().equals("A")) {
+			switch(mapping) {
+				// insertRoom.jsp
+				case "insertForm" : view =  rc.insertForm(request, response); break;
+				case "insert" : view =  rc.insert(request, response); break;
+				
+				case "updateListForm" : view =  rc.updateListForm(request, response); break;
+				case "updateRoomForm" : view =  rc.updateRoomForm(request, response); break;
+				case "updateRoom" : view =  rc.updateRoom(request, response); break;
+				
+				case "deleteListForm" : view = rc.deleteListForm(request, response); break;
+			}
 		
-		switch(mapping) {
-			// insertRoom.jsp
-			case "insertForm" : view =  rc.insertForm(request, response); break;
-			case "insert" : view =  rc.insert(request, response); break;
-			
-			case "updateListForm" : view =  rc.updateListForm(request, response); break;
-			case "updateRoomForm" : view =  rc.updateRoomForm(request, response); break;
-			case "updateRoom" : view =  rc.updateRoom(request, response); break;
-			
-			case "deleteListForm" : view = rc.deleteListForm(request, response); break;
 		}
-		
 		if(flag) {
 			request.getRequestDispatcher(view).forward(request, response);
